att functions create and update documents

+++ server/src/models/documento.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 type Documento struct {
 	gorm.Model
 	Nome      string `json:"nome"`
-	Ano       int16  `json:"ano"`
+	Ano       int    `json:"ano"`
 	Categoria string `json:"categoria"`
 	Arquivo   string `json:"arquivo"`
 }
